requirements.txt
fastapi
uvicorn
python-multipart
streamlit
requests
pillow
python-dotenv
torch
torchvision
pydantic
pydantic-settings
ultralytics
opencv-python-headless
